Use correct documented assertion method
I don't remember how I thought I came across `called_with`. It's not
a method. Not even back in 3.6.

The tests only passed, because `render_html` was returning a mock and
every method called on a mock is another mock and it looks like mocks
are truthy. I guess I never verified this assertion with a failing test.