Refactor mock and assertion
The previous assertion method was not suitable, because the Context
object is complex to set up, so we could not just create a new Context
from a dict. We would have to recreate all the layers that the Context
goes through to make the context objects compareable.

Thus, it seemed more desireable to compare the effective context by
checking the keys and values.

However, this also ran into issues. The context object is mutable and
values / layers are bing popped off as rendering happens. So to be able
to inspect the context as it was at the time of being passed to the
render_html method, we need a special custom Mock subclass that stores
copies of the values.